Convert contrib to Typescript
